Move Location to Domain
Step 15 puts ResolvedSymbol in Domain, and ResolvedSymbol::getDefinitionLocation
returns a Location, so Location moves too.
