+++ src/Resolution/ResolvedParameter.php
@@ -6,7 +6,7 @@ namespace Firehed\PhpLsp\Resolution;
 
 use Firehed\PhpLsp\Domain\ParameterInfo;
 use Firehed\PhpLsp\Domain\Type;
-use Firehed\PhpLsp\Index\Location;
+use Firehed\PhpLsp\Domain\Location;
 
 /**
  * A resolved parameter wrapping ParameterInfo.
